LayoutTests/imported/w3c/web-platform-tests/html/cross-origin-opener-policy/reporting/navigation-reporting/reporting-coop-navigated-opener.https-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: cancelled

Harness Error (TIMEOUT), message = null

TIMEOUT navigation-report-from-opener-navigation Test timed out
